[HUDI-7598] Remove duplicate methods in subclasses of HoodieSparkClientTestBase to enhance reusability
