+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/functional/RecordLevelIndexTestBase.scala
@@ -19,36 +19,25 @@ package org.apache.hudi.functional
 
 import org.apache.hudi.DataSourceWriteOptions
 import org.apache.hudi.DataSourceWriteOptions._
-import org.apache.hudi.client.SparkRDDWriteClient
-import org.apache.hudi.client.common.HoodieSparkEngineContext
-import org.apache.hudi.common.config.{HoodieMetadataConfig, TypedProperties}
-import org.apache.hudi.common.model._
+import org.apache.hudi.common.config.HoodieMetadataConfig
 import org.apache.hudi.common.table.{HoodieTableConfig, HoodieTableMetaClient}
-import org.apache.hudi.common.table.timeline.{HoodieInstant, HoodieInstantTimeGenerator, MetadataConversionUtils}
-import org.apache.hudi.common.testutils.HoodieTestUtils.INSTANT_FILE_NAME_GENERATOR
 import org.apache.hudi.common.testutils.RawTripTestPayload.recordsToStrings
 import org.apache.hudi.config.HoodieWriteConfig
 import org.apache.hudi.metadata.{HoodieBackedTableMetadata, HoodieTableMetadataUtil, MetadataPartitionType}
-import org.apache.hudi.storage.StoragePath
-import org.apache.hudi.testutils.HoodieSparkClientTestBase
-import org.apache.hudi.util.JavaConversions
+
 
 import org.apache.spark.sql._
-import org.apache.spark.sql.functions.{col, not}
 import org.junit.jupiter.api.Assertions.{assertEquals, assertFalse, assertTrue}
 import org.junit.jupiter.api._
 
 import java.util.concurrent.atomic.AtomicInteger
-import java.util.stream.Collectors
 
 import scala.collection.JavaConverters._
 import scala.collection.{JavaConverters, mutable}
 import scala.util.Using
 
-class RecordLevelIndexTestBase extends HoodieSparkClientTestBase {
-  var spark: SparkSession = _
-  var instantTime: AtomicInteger = _
-  val metadataOpts = Map(
+class RecordLevelIndexTestBase extends HoodieStatsIndexTestBase {
+  val metadataOpts: Map[String, String] = Map(
     HoodieMetadataConfig.ENABLE.key -> "true",
     HoodieMetadataConfig.RECORD_INDEX_ENABLE_PROP.key -> "true"
   )
@@ -81,8 +70,6 @@ class RecordLevelIndexTestBase extends HoodieSparkClientTestBase {
   ) ++ metadataOpts
 
   val commonOptsWithSecondaryIndexSITest = commonOptsNewTableSITest ++ secondaryIndexOpts
-  var mergedDfList: List[DataFrame] = List.empty
-  var metaClientReloaded = false
 
   @BeforeEach
   override def setUp() {
@@ -105,83 +92,6 @@ class RecordLevelIndexTestBase extends HoodieSparkClientTestBase {
   override def tearDown() = {
     cleanupFileSystem()
     cleanupSparkContexts()
-  }
-
-  protected def getLatestMetaClient(enforce: Boolean): HoodieTableMetaClient = {
-    val lastInsant = HoodieInstantTimeGenerator.getLastInstantTime
-    if (enforce || metaClient.getActiveTimeline.lastInstant().isEmpty
-      || metaClient.getActiveTimeline.lastInstant().get().requestedTime.compareTo(lastInsant) < 0) {
-      println("Reloaded timeline")
-      metaClient.reloadActiveTimeline()
-      metaClient
-    }
-    metaClient
-  }
-
-  protected def rollbackLastInstant(hudiOpts: Map[String, String]): HoodieInstant = {
-    val lastInstant = getLatestMetaClient(false).getActiveTimeline
-      .filter(JavaConversions.getPredicate(instant => instant.getAction != ActionType.rollback.name()))
-      .lastInstant().get()
-    if (getLatestCompactionInstant() != getLatestMetaClient(false).getActiveTimeline.lastInstant()
-      && lastInstant.getAction != ActionType.replacecommit.name()
-      && lastInstant.getAction != ActionType.clean.name()) {
-      mergedDfList = mergedDfList.take(mergedDfList.size - 1)
-    }
-    val writeConfig = getWriteConfig(hudiOpts)
-    new SparkRDDWriteClient(new HoodieSparkEngineContext(jsc), writeConfig)
-      .rollback(lastInstant.requestedTime)
-
-    if (lastInstant.getAction != ActionType.clean.name()) {
-      assertEquals(ActionType.rollback.name(), getLatestMetaClient(true).getActiveTimeline.lastInstant().get().getAction)
-    }
-    lastInstant
-  }
-
-  protected def executeFunctionNTimes[T](function0: Function0[T], n: Int): Unit = {
-    for (i <- 1 to n) {
-      function0.apply()
-    }
-  }
-
-  protected def deleteLastCompletedCommitFromDataAndMetadataTimeline(hudiOpts: Map[String, String]): Unit = {
-    val writeConfig = getWriteConfig(hudiOpts)
-    val lastInstant = getHoodieTable(getLatestMetaClient(false), writeConfig).getCompletedCommitsTimeline.lastInstant().get()
-    val metadataTableMetaClient = getHoodieTable(metaClient, writeConfig).getMetadataTable.asInstanceOf[HoodieBackedTableMetadata].getMetadataMetaClient
-    val metadataTableLastInstant = metadataTableMetaClient.getCommitsTimeline.lastInstant().get()
-    assertTrue(storage.deleteFile(new StoragePath(metaClient.getTimelinePath, INSTANT_FILE_NAME_GENERATOR.getFileName(lastInstant))))
-    assertTrue(storage.deleteFile(new StoragePath(
-      metadataTableMetaClient.getTimelinePath, INSTANT_FILE_NAME_GENERATOR.getFileName(metadataTableLastInstant))))
-    mergedDfList = mergedDfList.take(mergedDfList.size - 1)
-  }
-
-  protected def deleteLastCompletedCommitFromTimeline(hudiOpts: Map[String, String]): Unit = {
-    val writeConfig = getWriteConfig(hudiOpts)
-    val lastInstant = getHoodieTable(getLatestMetaClient(false), writeConfig).getCompletedCommitsTimeline.lastInstant().get()
-    assertTrue(storage.deleteFile(new StoragePath(metaClient.getTimelinePath, INSTANT_FILE_NAME_GENERATOR.getFileName(lastInstant))))
-    mergedDfList = mergedDfList.take(mergedDfList.size - 1)
-  }
-
-  protected def getMetadataMetaClient(hudiOpts: Map[String, String]): HoodieTableMetaClient = {
-    getHoodieTable(metaClient, getWriteConfig(hudiOpts)).getMetadataTable.asInstanceOf[HoodieBackedTableMetadata].getMetadataMetaClient
-  }
-
-  protected def getLatestCompactionInstant(): org.apache.hudi.common.util.Option[HoodieInstant] = {
-    getLatestMetaClient(false).getActiveTimeline
-      .filter(JavaConversions.getPredicate(s => Option(
-        try {
-          val commitMetadata = MetadataConversionUtils.getHoodieCommitMetadata(metaClient, s)
-            .orElse(new HoodieCommitMetadata())
-          commitMetadata
-        } catch {
-          case _: Exception => new HoodieCommitMetadata()
-        })
-        .map(c => c.getOperationType == WriteOperationType.COMPACT)
-        .get))
-      .lastInstant()
-  }
-
-  protected def getLatestClusteringInstant(): org.apache.hudi.common.util.Option[HoodieInstant] = {
-    getLatestMetaClient(false).getActiveTimeline.getCompletedReplaceTimeline.lastInstant()
   }
 
   protected def doWriteAndValidateDataAndRecordIndex(hudiOpts: Map[String, String],
@@ -229,60 +139,7 @@ class RecordLevelIndexTestBase extends HoodieSparkClientTestBase {
   }
 
   protected def calculateMergedDf(latestBatchDf: DataFrame, operation: String): DataFrame = {
-    calculateMergedDf(latestBatchDf, operation, false)
-  }
-
-  /**
-   * @return [[DataFrame]] that should not exist as of the latest instant; used for non-existence validation.
-   */
-  protected def calculateMergedDf(latestBatchDf: DataFrame, operation: String, globalIndexEnableUpdatePartitions: Boolean): DataFrame = {
-    val prevDfOpt = mergedDfList.lastOption
-    if (prevDfOpt.isEmpty) {
-      mergedDfList = mergedDfList :+ latestBatchDf
-      sparkSession.emptyDataFrame
-    } else {
-      if (operation == INSERT_OVERWRITE_TABLE_OPERATION_OPT_VAL) {
-        mergedDfList = mergedDfList :+ latestBatchDf
-        // after insert_overwrite_table, all previous snapshot's records should be deleted from RLI
-        prevDfOpt.get
-      } else if (operation == INSERT_OVERWRITE_OPERATION_OPT_VAL) {
-        val overwrittenPartitions = latestBatchDf.select("partition")
-          .collectAsList().stream().map[String](JavaConversions.getFunction[Row, String](r => r.getString(0))).collect(Collectors.toList[String])
-        val prevDf = prevDfOpt.get
-        val latestSnapshot = prevDf
-          .filter(not(col("partition").isInCollection(overwrittenPartitions)))
-          .union(latestBatchDf)
-        mergedDfList = mergedDfList :+ latestSnapshot
-
-        // after insert_overwrite (partition), all records in the overwritten partitions should be deleted from RLI
-        prevDf.filter(col("partition").isInCollection(overwrittenPartitions))
-      } else {
-        val prevDf = prevDfOpt.get
-        if (globalIndexEnableUpdatePartitions) {
-          val prevDfOld = prevDf.join(latestBatchDf, prevDf("_row_key") === latestBatchDf("_row_key"), "leftanti")
-          val latestSnapshot = prevDfOld.union(latestBatchDf)
-          mergedDfList = mergedDfList :+ latestSnapshot
-        } else {
-          val prevDfOld = prevDf.join(latestBatchDf, prevDf("_row_key") === latestBatchDf("_row_key")
-            && prevDf("partition") === latestBatchDf("partition"), "leftanti")
-          val latestSnapshot = prevDfOld.union(latestBatchDf)
-          mergedDfList = mergedDfList :+ latestSnapshot
-        }
-        sparkSession.emptyDataFrame
-      }
-    }
-  }
-
-  private def getInstantTime(): String = {
-    String.format("%03d", new Integer(instantTime.incrementAndGet()))
-  }
-
-  protected def getWriteConfig(hudiOpts: Map[String, String]): HoodieWriteConfig = {
-    val props = TypedProperties.fromMap(JavaConverters.mapAsJavaMapConverter(hudiOpts).asJava)
-    HoodieWriteConfig.newBuilder()
-      .withProps(props)
-      .withPath(basePath)
-      .build()
+    calculateMergedDf(latestBatchDf, operation, globalIndexEnableUpdatePartitions = false)
   }
 
   protected def getFileGroupCountForRecordIndex(writeConfig: HoodieWriteConfig): Long = {
